+++ VisualPinball.Unity/VisualPinball.Unity.Editor/Inspectors/MechSoundsInspector.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d745160357eec104ab1beb6bd2ebdd3c
+guid: 8dca340821f92c74e8cf97cd3fff29df
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
